Refactor: 리뷰 페이지를 모달로 변경
