--- client/src/pages/review-register/ReviewRegister.style.ts
@@ -1,146 +1,0 @@
-import { styled } from 'styled-components';
-
-export default {
-  Heading1: styled.h1`
-    margin-top: 20px;
-    margin-left: 15px;
-    margin-right: auto;
-    color: white;
-    font-size: var(--heading1-font-size);
-    font-weight: var(--heading1-font-weight);
-    line-height: var(--heading1-line-height);
-  `,
-  Heading3: styled.h3`
-    margin-top: 20px;
-    margin-bottom: 8px;
-    margin-right: auto;
-    margin-left: 15px;
-    color: white;
-    font-size: var(--heading3-font-size);
-    font-weight: var(--heading3-font-weight);
-    line-height: var(--heading3-line-height);
-  `,
-  Container: styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #1e1e1e;
-  `,
-  Main: styled.main`
-    width: 390px;
-    min-height: calc(100vh - 50px);
-    background-color: var(--theme-background-color);
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    padding-bottom: 70px;
-    & h1 {
-      margin-right: auto;
-      margin-left: 15px;
-      margin-top: 20px;
-    }
-  `,
-  CategoryContainer: styled.div`
-    margin-top: 20px;
-    display: flex;
-    min-width: 360px;
-    justify-content: flex-start;
-    & > button:not(:first-child) {
-      margin-left: 13px;
-    }
-  `,
-  Summary: styled.div`
-    margin-top: 20px;
-    width: 360px;
-    gap: 8px;
-    display: flex;
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    & p {
-      color: white;
-      font-size: var(--p-small-medium-font-size);
-      font-weight: var(--p-small-medium-font-weight);
-      line-height: var(--p-small-medium-line-height);
-    }
-  `,
-  FileInput: styled.input`
-    display: none;
-  `,
-  Poster: styled.img`
-    margin-top: 8px;
-    width: 170px;
-    height: 210px;
-    background-color: gray;
-    cursor: pointer;
-    object-fit: scale-down;
-    &[src] {
-      background-color: transparent;
-    }
-  `,
-  // align-items: flex-start 왼쪽 정렬, flex-end 오른쪽 정렬
-  Form: styled.form`
-    width: 360px;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-    & label {
-      color: white;
-      font-size: var(--p-small-medium-font-size);
-      font-weight: var(--p-small-medium-font-weight);
-      line-height: var(--p-small-medium-line-height);
-    }
-    & span {
-      color: var(--font-highlight-color);
-      font-size: var(--p-small-medium-font-size);
-      font-weight: var(--p-small-medium-font-weight);
-      line-height: var(--p-small-medium-line-height);
-    }
-  `,
-  TextareaContainer: styled.textarea`
-    width: 360px;
-    min-height: 150px;
-    font-size: var(--p-small-medium-font-size);
-    font-weight: var(--p-small-medium-font-weight);
-    line-height: var(--p-small-medium-line-height);
-    outline: none;
-    resize: vertical;
-    padding: 1px;
-  `,
-  TitleButtonFlex: styled.div`
-    width: 360px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    & > h3 {
-      margin-left: 0;
-    }
-  `,
-  Map: styled.div`
-    width: 360px;
-    height: 200px;
-    background-color: gray;
-  `,
-  Image: styled.div`
-    width: 360px;
-    height: 210px;
-    background-color: gray;
-    margin-bottom: 20px;
-  `,
-  BottomStickyContainer: styled.div`
-    width: 390px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: fixed;
-    top: auto;
-    bottom: 70px;
-    & button {
-      z-index: 1;
-    }
-  `,
-};
